backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
psycopg2-binary==2.9.10
xlrd==2.0.1
openpyxl==3.0.9
python-dotenv==0.19.0
werkzeug==2.0.1
sqlalchemy==1.4.23 